A feature for adding/editing answers of a question
On a page with a question a list of answers is displayed with a form to
add a new answer. A user can add, edit and delete his own answers. He
can only read answers of others.

An answer form and edit/delete/cancel links trigger AJAX requests.

Rails UJS and jQuery javascript libraries are added to enable server
generated javascript responses. A :js MIME type is also configured.