Add a feature for managing user’s questions
A user can post a question that is displayed on a list of questions.
Later we’ll add a feature for posting answers to a question.

A `questions` database table has fields:
- `title`: a short summary of a question
- `body`: a larger description of a question
- `slug`: a slug automatically generated from a title intended to be
  displayed in an url
- `user_id`: an author of a question
- `tags`: a list of tags relevant to a question
- `updated_at`
- `inserted_at`

When a user posts a question, a title and a body are required fields,
a field with tags can be left blank. Fields `slug` and `user_id` are
internally generated and assigned.

We add a custom Ecto Type called `Experts.EctoTag` that knows how to
cast a string of tags separated with a comma to a Postgres array column.

Add a `slugify` hex package for generating an url slug from a title.

Disable credo warnings for TODO comments.