Make `/questions` and `/questions/:id` pages public
Any visitor can see a list of questions and a page with a question and
a list of answers.

Only logged in users can post questions and answers. Display write an
answer or write a question forms to logged in users only.

If a user is not logged in, display links to login or to become a
member.

An answer edit/delete links are always hidden from public visitors.