Raise an error if an unauthorized user tries to modify a question
We removes redirects and flash message warnings about permissions
because they are not necessary.

If a user who is not the author of a question tries to edit, update or
delete a question, raise an error. If an error happens it means that
there is probably a bug in views - links/forms to perform these actions
should not be visible to unauthorized users. We need to be notified
about that instead of displaying a flash message to a user.

In this commit we also add some documentation about a purpose
of a question controller.