Add a user profile edit form, add a field `name` as required
A question/an answer has to display the name of a user who posted it.
We make a `name` as a required field when a user registers. We also add
a link to a page `registration/edit` where a user can edit account
details. The page is handled by Pow library.